backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pandapower>=2.13.0
pdfminer.six>=20221105
PyMuPDF>=1.23.8
pdfminer.six>=20221105
